[step2] feat: 로또의 일치하는 개수 확인 기능 추가
