+++ README.md
@@ -32,7 +32,7 @@
 - [x] '지난주 당첨 번호를 입력해 주세요.' 메시지를 출력합니다. - InputView
 - [x] 지난주 당첨 번호를 입력받습니다. (6개 입력, 범위: 1~45) - InputView
  
-- [ ] 각 로또마다 일치하는 개수를 계산합니다. - Lotto
+- [x] 각 로또마다 일치하는 개수를 계산합니다. - Lotto
 - [ ] 모든 로또의 일치하는 개수를 통한 통계를 냅니다. - LottoStore? Person?
 - [ ] 전체 로또에 대한 수익률을 계산합니다. - LottoStore? Person?
 - [ ] 당첨 통계와 총 수익률을 출력합니다. - OutputView
