+++ src/main/kotlin/lotto/Lotto.kt
@@ -2,6 +2,7 @@ package lotto
 
 import lotto.view.output.OutputView
 
-interface Lotto {
+interface Lotto<T, V> {
     fun renderLotto(outputView: OutputView)
+    fun checkWinning(winningValue: T): V
 }
