+++ src/main/kotlin/lotto/LottoStore.kt
@@ -1,6 +1,6 @@
 package lotto
 
 interface LottoStore {
-    fun purchase(count: Int): List<Lotto>
-    fun makeLotto(): Lotto
+    fun purchase(count: Int): List<Lotto<*, *>>
+    fun makeLotto(): Lotto<*, *>
 }
