+++ src/main/kotlin/lotto/SixFortyFiveLotto.kt
@@ -1,31 +1,14 @@
 package lotto
 
 import lotto.view.output.OutputView
-import kotlin.random.Random
 
-class SixFortyFiveLotto : Lotto {
-    val numbers: List<Int>
-
-    init {
-        numbers = generateLottoNumber()
-    }
+class SixFortyFiveLotto(val numbers: List<Int>) : Lotto<List<Int>, Int> {
 
     override fun renderLotto(outputView: OutputView) {
         outputView.renderMessage()
     }
 
-    private fun generateLottoNumber(): List<Int> {
-        return (1..LOTTO_NUMBER_COUNT).map {
-            Random.nextInt(
-                LOTTO_NUMBER_RANGE_START,
-                LOTTO_NUMBER_RANGE_END,
-            )
-        }.sorted()
-    }
-
-    companion object {
-        const val LOTTO_NUMBER_COUNT = 6
-        const val LOTTO_NUMBER_RANGE_START = 1
-        const val LOTTO_NUMBER_RANGE_END = 45
+    override fun checkWinning(winningValue: List<Int>): Int {
+        return numbers.count { winningValue.contains(it) }
     }
 }
